+++ lib/Settings/Personal.php
@@ -7,6 +7,7 @@ namespace OCA\DocSort\Settings;
 use OCA\DocSort\AppInfo\Application;
 use OCA\DocSort\Service\DefaultRules;
 use OCA\DocSort\Service\Organizer;
+use OCA\DocSort\Service\PythonEnv;
 use OCA\DocSort\Service\Settings;
 use OCP\AppFramework\Http\TemplateResponse;
 use OCP\AppFramework\Services\IInitialState;
@@ -15,7 +16,7 @@ use OCP\Settings\ISettings;
 
 final class Personal implements ISettings
 {
-    public function __construct(private IInitialState $state, private Settings $settings, private Organizer $organizer, private IUserSession $session) {}
+    public function __construct(private IInitialState $state, private Settings $settings, private Organizer $organizer, private IUserSession $session, private PythonEnv $env) {}
 
     public function getForm(): TemplateResponse
     {
@@ -28,6 +29,7 @@ final class Personal implements ISettings
         $this->state->provideInitialState('history', $this->organizer->history($uid));
         $this->state->provideInitialState('kinds', $names);
         $this->state->provideInitialState('categories', DefaultRules::CATEGORIES);
+        $this->state->provideInitialState('readerReady', $this->env->isInstalled() || $this->settings->hasExternalReader());
 
         return new TemplateResponse(Application::APP_ID, 'personal', [], '');
     }